refactor; efficient mutexes; widgets react to CAN data
Widgets will now at least react to incoming frame data and update their displays. Introduced the concept of "Channels" to widgets, which indicate the type of signal being captured. Later, these will be defined as hard references (such as speed_data and oil_pressure_data). Once signal decoding is developed, the plan is NOT to memcpy the entire message frame into each signal, but it's good for PoC work.