restructure; add widget skins by conf; fragment widgets
Widgets can now select skins! Since the customization for these is rather limitless, it's good to separate each widget type into a folder as this update does. The purpose of the generic widget file now becomes (1) widget initialization and channel selection, and (2) to source all sub-components together as one compilation unit. NOTE THAT A WIDGET IS STILL TIED TO A REQUIRED SET AND ORDERING OF CHANNELS. SKINS SHOULD NOT MODIFY THIS.