MILESTONE: my needle meter widget is now working to get SPEED input from ICSim!
Represents a huge leap in the project. I can now render a widget (~100us total per frame) which is controlled from an entirely external CAN application, and can prove asynchronous widget rendering works too. This project is seriously speedy, can't wait to make it as lightweight and flexible as possible.